put stock code into seperate app
