+++ fintrack/management/commands/seed_data.py
@@ -1,6 +1,6 @@
 from django.core.management import BaseCommand
 from fintrack.management.commands.sub_methods import countries, exchanges, fix_empty_sector_industry, index
-from sector import sector_data
+from sector.services import sector_data
 from industry import industry_data
 from company.services import company_data
