+++ fintrack/urls.py
@@ -13,4 +13,5 @@ urlpatterns = [
     path('api/v1/industry/', include('industry.urls')),
     path('api/v1/exchange/', include('exchange.urls')),
     path('api/v1/company/', include('company.urls')),
+    path('api/v1/stock/', include('stock.urls')),
 ]
