+++ fintrack_be/admin/__init__.py
@@ -1,5 +1,3 @@
-from .stock_admin import *
-from .stock_price_admin import *
 from .index_admin import *
 
 admin.site.site_header = "FinTrack"
