--- fintrack_be/functions/stock/__init__.py
@@ -1,1 +1,0 @@
-from .stock_data import *
