+++ fintrack_be/models/__init__.py
@@ -1,5 +1,3 @@
-from .stock import Stock
-from .stock_price import StockPriceData
 from .index import Index
 from .index_constituents import IndexConstituents
 from .index_price import IndexPriceData
