+++ fintrack_be/tasks/__init__.py
@@ -1,2 +1,1 @@
-from .stock_tasks import *
 from .email_tasks import *
