+++ industry/serializers/__init__.py
@@ -1,0 +1,1 @@
+from .industry_serializer import *
