+++ industry/views/industry_views.py
@@ -8,10 +8,10 @@ from rest_framework.views import APIView
 
 from industry.models import Industry
 from company.models import Company
-from fintrack_be.models import Stock, StockPriceData
-from industry.serializers.industry_serializer import IndustrySerializer
-from company.serializers.company_serializer import CompanySerializer
-from fintrack_be.serializers.stock_serializer import BasicStockSerializer
+from industry.serializers import IndustrySerializer
+from company.serializers import CompanySerializer
+from stock.models import Stock, StockPriceData
+from stock.serializers import BasicStockSerializer
 from fintrack_be.permissions import IsVerified
