+++ sector/views/sector_views.py
@@ -7,12 +7,12 @@ from rest_framework.response import Response
 from rest_framework.views import APIView
 
 from sector.models import Sector
+from sector.serializers import SectorSerializer
 from industry.models import Industry
 from company.models import Company
-from fintrack_be.models import Stock, StockPriceData
-from sector.serializers.sector_serializer import SectorSerializer
-from company.serializers.company_serializer import CompanySerializer
-from fintrack_be.serializers.stock_serializer import BasicStockSerializer
+from company.serializers import CompanySerializer
+from stock.models import Stock, StockPriceData
+from stock.serializers import BasicStockSerializer
 from fintrack_be.permissions import IsVerified
