+++ stock/admin/__init__.py
@@ -1,0 +1,2 @@
+from .stock_admin import *
+from .stock_price_admin import *
