+++ stock/models/__init__.py
@@ -1,0 +1,2 @@
+from .stock import Stock
+from .stock_price import *
