+++ stock/serializers/stock_serializer.py
@@ -1,5 +1,5 @@
 from rest_framework import serializers
-from fintrack_be.models import Stock, StockPriceData
+from stock.models import Stock
 
 
 class StockSerializer(serializers.ModelSerializer):
@@ -26,34 +26,3 @@ class StockSerializer(serializers.ModelSerializer):
                   'change',
                   'change_perc',
                   'price_date_utc')
-
-
-class BasicStockSerializer(serializers.ModelSerializer):
-    company_name = serializers.CharField(source='company.short_name')
-    exchange_symbol = serializers.CharField(source='exchange.symbol')
-    industry_name = serializers.CharField(source='company.industry.name')
-
-    class Meta:
-        model = Stock
-        fields = ('ticker',
-                  'name',
-                  'exchange_symbol',
-                  'company_name',
-                  'industry_name')
-
-
-class StockPriceDataSerializer(serializers.ModelSerializer):
-    stock_symbol = serializers.CharField(source='stock.ticker')
-
-    class Meta:
-        model = StockPriceData
-        fields = ('timestamp',
-                  'timestamp_in_market_time',
-                  'stock_symbol',
-                  'high',
-                  'low',
-                  'open',
-                  'close',
-                  'volume',
-                  'change',
-                  'change_perc')
