+++ stock/tasks/stock_tasks.py
@@ -8,7 +8,7 @@ def get_latest_stock_data(ticker):
     Method that gets the latest data for a Stock
     :param ticker: Ticker of Stock to get data for
     """
-    from fintrack_be.functions.stock.stock_data import get_stock_data, stock_price_data_df_to_model
+    from stock.services import get_stock_data, stock_price_data_df_to_model
 
     df = get_stock_data(ticker, period='1d', interval='1m')
     stock_price_data_df_to_model(df)
@@ -21,7 +21,7 @@ def get_day_stock_data(ticker):
     Method that that gets the day data for a Stock
     :param ticker: Stock ticker to get data for
     """
-    from fintrack_be.functions.stock.stock_data import get_stock_data, stock_price_data_df_to_model
+    from stock.services import get_stock_data, stock_price_data_df_to_model
 
     df = get_stock_data(ticker, period='1d', interval='1d')
     stock_price_data_df_to_model(df)
@@ -34,7 +34,7 @@ def get_bulk_day_stock_data(ticker):
     Method that that gets the day data for a Stock
     :param ticker: Stock symbool to get data for
     """
-    from fintrack_be.functions.stock.stock_data import get_stock_data, bulk_stock_price_data_to_model
+    from stock.services import get_stock_data, bulk_stock_price_data_to_model
 
     df = get_stock_data(ticker, period='1y', interval='1d')
     bulk_stock_price_data_to_model(df)
