+++ stock/urls/__init__.py
@@ -1,0 +1,1 @@
+from .stock_urls import *
